requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery~=4.4.1
certifi==2020.6.20
chardet==3.0.4
Deprecated==1.2.10
# Django==3.1.1
django-phone-verify==2.0.0
django-phonenumber-field==5.0.0
djangorestframework~=3.11.0
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
# mysqlclient==2.0.1
pyaes==1.6.1
pyasn1==0.4.8
PyJWT==1.7.1
PySocks==1.7.1
python-dotenv~=0.15.0
pytz~=2020.4
redis==3.5.3
redis-server==5.0.7
requests~=2.22.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
Telethon~=1.17.5
twilio==6.45.2
urllib3==1.25.11
vine==1.3.0
wrapt==1.12.1
psycopg2-binary==2.8.6
BeautifulSoup4==4.9.3
bs4==2.2.1
facebook-scraper==0.2.51